remove duplicate macro rocksdb_enum_mod
